pyproject.toml
[project]
name = "cupido"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bitsandbytes>=0.46.1",
    "datasets>=4.0.0",
    "flash-attn==2.8.2",
    "grobid-client-python>=0.0.11",
    "llamore>=0.2.0",
    "peft>=0.16.0",
    "pillow>=11.3.0",
    "pydantic-settings>=2.10.1",
    "pymupdf>=1.26.3",
    "qwen-vl-utils>=0.0.11",
    "torchvision>=0.22.1",
    "transformers>=4.53.1",
    "trl>=0.19.1",
    "wandb>=0.21.0",
]

[dependency-groups]
dev = [
    "ipywidgets>=8.1.7",
    "jupyterlab>=4.4.4",
]
